move overlays to io.visualization
